+++ src/main/java/ru/otus/library/app/genre/dto/response/DtoGetGenreResponse.java
@@ -2,23 +2,23 @@ package ru.otus.library.app.genre.dto.response;
 
 public class DtoGetGenreResponse {
 
-    private Integer id;
+    private Long id;
 
     private String name;
 
     public DtoGetGenreResponse() {
     }
 
-    public DtoGetGenreResponse(Integer id, String name) {
+    public DtoGetGenreResponse(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
