+++ src/main/java/ru/otus/library/domain/repositories/BookRepositoryJdbc.java
@@ -63,21 +63,21 @@ public class BookRepositoryJdbc implements BookRepository {
                 "          LEFT JOIN book_author ba on b.id = ba.book_id\n" +
                 "          LEFT JOIN author a on ba.author_id = a.id";
 
-        Map<Integer, BookResultSet> bookResultSetMap = new HashMap<>();
+        Map<Long, BookResultSet> bookResultSetMap = new HashMap<>();
 
         namedParamsJdbcOperations.query(query, rs -> {
-            Integer bookId = rs.getInt("book_id");
+            Long bookId = rs.getLong("book_id");
 
             BookResultSet bookResultSet = bookResultSetMap.getOrDefault(bookId, new BookResultSet());
 
-            Integer authorId = rs.getInt("author_id");
+            Long authorId = rs.getLong("author_id");
             if (authorId > 0) {
                 String firstName = rs.getString("author_first_name");
                 String lastName = rs.getString("author_last_name");
                 bookResultSet.authors.add(new DbAuthor(authorId, firstName, lastName));
             }
 
-            Integer genreId = rs.getInt("genre_id");
+            Long genreId = rs.getLong("genre_id");
             if (genreId > 0) {
                 String name = rs.getString("genre_name");
                 bookResultSet.genres.add(new DbGenre(genreId, name));
@@ -118,7 +118,7 @@ public class BookRepositoryJdbc implements BookRepository {
         KeyHolder keyHolder = new GeneratedKeyHolder();
 
         createOrUpdate(keyHolder, book);
-        book.setId((Integer) keyHolder.getKey());
+        book.setId(keyHolder.getKey().longValue());
 
         batchUpdateGenres(book.getId(), book.getGenres());
         batchUpdateAuthors(book.getId(), book.getAuthors());
@@ -136,20 +136,20 @@ public class BookRepositoryJdbc implements BookRepository {
         namedParamsJdbcOperations.getJdbcOperations().update(con -> {
             PreparedStatement preparedStatement = con.prepareStatement(getCreateOrUpdateSql(book), new String[]{ "id" });
             preparedStatement.setString(1, book.getTitle());
-            if (book.getId() != null) { preparedStatement.setInt(2, book.getId()); }
+            if (book.getId() != null) { preparedStatement.setLong(2, book.getId()); }
             return preparedStatement;
         }, keyHolder);
     }
 
-    private void cleanBookGenresByBookId(Integer bookId) {
+    private void cleanBookGenresByBookId(Long bookId) {
         namedParamsJdbcOperations.getJdbcOperations().update("DELETE FROM book_genre WHERE book_id = ?", bookId);
     }
 
-    private void cleanBookAuthorsByBookId(Integer bookId) {
+    private void cleanBookAuthorsByBookId(Long bookId) {
         namedParamsJdbcOperations.getJdbcOperations().update("DELETE FROM book_author WHERE book_id = ?", bookId);
     }
 
-    private void batchUpdateGenres(Integer bookId, final List<DbGenre> genres) {
+    private void batchUpdateGenres(Long bookId, final List<DbGenre> genres) {
 
         cleanBookGenresByBookId(bookId);
 
@@ -157,8 +157,8 @@ public class BookRepositoryJdbc implements BookRepository {
                 "insert into book_genre (book_id, genre_id) values (?, ?)",
                 new BatchPreparedStatementSetter() {
                     public void setValues(PreparedStatement ps, int i) throws SQLException {
-                        ps.setInt(1, bookId);
-                        ps.setInt(2, genres.get(i).getId());
+                        ps.setLong(1, bookId);
+                        ps.setLong(2, genres.get(i).getId());
                     }
 
                     public int getBatchSize() {
@@ -167,7 +167,7 @@ public class BookRepositoryJdbc implements BookRepository {
                 } );
     }
 
-    private void batchUpdateAuthors(Integer bookId, final List<DbAuthor> authors) {
+    private void batchUpdateAuthors(Long bookId, final List<DbAuthor> authors) {
 
         cleanBookAuthorsByBookId(bookId);
 
@@ -176,8 +176,8 @@ public class BookRepositoryJdbc implements BookRepository {
                 new BatchPreparedStatementSetter() {
                     @Override
                     public void setValues(PreparedStatement ps, int i) throws SQLException {
-                        ps.setInt(1, bookId);
-                        ps.setInt(2, authors.get(i).getId());
+                        ps.setLong(1, bookId);
+                        ps.setLong(2, authors.get(i).getId());
                     }
 
                     @Override
@@ -191,7 +191,7 @@ public class BookRepositoryJdbc implements BookRepository {
     private static class BookRowMapper implements RowMapper<DbBook> {
         @Override
         public DbBook mapRow(ResultSet rs, int rowNum) throws SQLException {
-            Integer id = rs.getInt("id");
+            Long id = rs.getLong("id");
             String title = rs.getString("title");
             return new DbBook(id, title);
         }
